Server/requirements.txt
python-dotenv
supabase
Jinja2
psycopg2